// src/modules/role/domain/enums/ability-code.enum.ts
export enum EAbilityCode {
  ADMIN = 'ADMIN',
  COMPANIES = 'COMPANIES',
  USERS = 'USERS',
  ROLES = 'ROLES',
  HAIRCUTS = 'HAIRCUTS',
  HOURS = 'HOURS',
  SCHEDULINGS = 'SCHEDULINGS',
  WORKING_DAYS = 'WORKING_DAYS',
  HOLIDAYS = 'HOLIDAYS',
  ASSESSMENTS = 'ASSESSMENTS',
  HISTORIES = 'HISTORIES',
}

export enum EAbilityCodeTranslate {
  ADMIN = 'Admin',
  COMPANIES = 'Empresas',
  USERS = 'Usuários',
  ROLES = 'Perfis',
  HAIRCUTS = 'Cortes',
  HOURS = 'Horários',
  SCHEDULINGS = 'Agendamentos',
  WORKING_DAYS = 'Dias de trabalho',
  HOLIDAYS = 'Feriados',
  ASSESSMENTS = 'Avaliações',
  HISTORIES = 'Históricos',
}